requirements.txt
networkx==2.3
jsonpath-rw>=1.4.0
requests>=2.21.0
graphviz>=0.11.1
aiohttp
pandas
pyyaml
